+++ heronpy/dsl/streamlet.py
@@ -17,7 +17,7 @@ from abc import abstractmethod
 
 from heronpy.api.topology import TopologyBuilder
 
-from heronpy.dsl.dslboltbase import DslBoltBase
+from heronpy.dsl.impl.dslboltbase import DslBoltBase
 
 # pylint: disable=too-many-instance-attributes
 class Streamlet(object):
@@ -55,7 +55,7 @@ class Streamlet(object):
   def map(self, map_function):
     """Return a new Streamlet by applying map_function to each element of this Streamlet.
     """
-    from heronpy.dsl.mapbolt import MapStreamlet
+    from heronpy.dsl.impl.mapbolt import MapStreamlet
     map_streamlet = MapStreamlet(map_function, self)
     self._add_child(map_streamlet)
     return map_streamlet
@@ -64,7 +64,7 @@ class Streamlet(object):
     """Return a new Streamlet by applying map_function to each element of this Streamlet
        and flattening the result
     """
-    from heronpy.dsl.flatmapbolt import FlatMapStreamlet
+    from heronpy.dsl.impl.flatmapbolt import FlatMapStreamlet
     fm_streamlet = FlatMapStreamlet(flatmap_function, self)
     self._add_child(fm_streamlet)
     return fm_streamlet
@@ -72,7 +72,7 @@ class Streamlet(object):
   def filter(self, filter_function):
     """Return a new Streamlet containing only the elements that satisfy filter_function
     """
-    from heronpy.dsl.filterbolt import FilterStreamlet
+    from heronpy.dsl.impl.filterbolt import FilterStreamlet
     filter_streamlet = FilterStreamlet(filter_function, self)
     self._add_child(filter_streamlet)
     return filter_streamlet
@@ -82,7 +82,7 @@ class Streamlet(object):
     num_partitions partitions. Note that this is different from num_partitions(n) in
     that new streamlet will be created by the repartition call.
     """
-    from heronpy.dsl.repartitionbolt import RepartitionStreamlet
+    from heronpy.dsl.impl.repartitionbolt import RepartitionStreamlet
     repartition_streamlet = RepartitionStreamlet(parallelism, self)
     self._add_child(repartition_streamlet)
     return repartition_streamlet
@@ -95,7 +95,7 @@ class Streamlet(object):
     It could also return a list of partitions if it wants to send it to multiple
     partitions.
     """
-    from heronpy.dsl.repartitionbolt import RepartitionStreamlet
+    from heronpy.dsl.impl.repartitionbolt import RepartitionStreamlet
     repartition_streamlet = RepartitionStreamlet(num_partitions, repartition_function, self)
     self._add_child(repartition_streamlet)
     return repartition_streamlet
@@ -115,7 +115,7 @@ class Streamlet(object):
       reduce_function takes two element at one time and reduces them to one element that
       is used in the subsequent operations.
     """
-    from heronpy.dsl.reducebywindowbolt import ReduceByWindowBolt
+    from heronpy.dsl.impl.reducebywindowbolt import ReduceByWindowBolt
     reduce_streamlet = ReduceByWindowBolt(window_config, reduce_function, self)
     self._add_child(reduce_streamlet)
     return reduce_streamlet
@@ -123,7 +123,7 @@ class Streamlet(object):
   def union(self, other_streamlet):
     """Returns a new Streamlet that consists of elements of both this and other_streamlet
     """
-    from heronpy.dsl.unionbolt import UnionBolt
+    from heronpy.dsl.impl.unionbolt import UnionBolt
     union_streamlet = UnionBolt(self, other_streamlet)
     self._add_child(union_streamlet)
     other_streamlet._add_child(union_streamlet)
@@ -135,7 +135,7 @@ class Streamlet(object):
     Before starting to cycle over the Streamlet, the open function of the transform_function is
     called. This allows the transform_function to do any kind of initialization/loading, etc.
     """
-    from heronpy.dsl.transformbolt import TransformStreamlet
+    from heronpy.dsl.impl.transformbolt import TransformStreamlet
     transform_streamlet = TransformStreamlet(transform_function, self)
     self._add_child(transform_streamlet)
     return transform_streamlet
@@ -143,7 +143,7 @@ class Streamlet(object):
   def log(self):
     """Logs all elements of this streamlet. This returns nothing
     """
-    from heronpy.dsl.logbolt import LogStreamlet
+    from heronpy.dsl.impl.logbolt import LogStreamlet
     log_streamlet = LogStreamlet(self)
     self._add_child(log_streamlet)
     return
@@ -151,7 +151,7 @@ class Streamlet(object):
   def to_sink(self, consumer_function):
     """Calls consumer_function for each element of this streamlet. This function returns nothing
     """
-    from heronpy.dsl.sinkbolt import SinkStreamlet
+    from heronpy.dsl.impl.sinkbolt import SinkStreamlet
     sink_streamlet = SinkStreamlet(consumer_function, self)
     self._add_child(sink_streamlet)
     return
@@ -159,7 +159,7 @@ class Streamlet(object):
   def join(self, join_streamlet, window_config, join_function):
     """Return a new Streamlet by joining join_streamlet with this streamlet
     """
-    from heronpy.dsl.joinbolt import JoinStreamlet
+    from heronpy.dsl.impl.joinbolt import JoinStreamlet
     join_streamlet = JoinStreamlet(JoinStreamlet.INNER, window_config,
                                    join_function, self, join_streamlet)
     self._add_child(join_streamlet)
@@ -169,7 +169,7 @@ class Streamlet(object):
   def outer_join(self, join_streamlet, window_config, join_function):
     """Return a new Streamlet by outer joining join_streamlet with this streamlet
     """
-    from heronpy.dsl.joinbolt import JoinStreamlet
+    from heronpy.dsl.impl.joinbolt import JoinStreamlet
     join_streamlet = JoinStreamlet(JoinStreamlet.OUTER, window_config,
                                    join_function, self, join_streamlet)
     self._add_child(join_streamlet)
@@ -179,7 +179,7 @@ class Streamlet(object):
   def left_join(self, join_streamlet, window_config, join_function):
     """Return a new Streamlet by left joining join_streamlet with this streamlet
     """
-    from heronpy.dsl.joinbolt import JoinStreamlet
+    from heronpy.dsl.impl.joinbolt import JoinStreamlet
     join_streamlet = JoinStreamlet(JoinStreamlet.LEFT, window_config,
                                    join_function, self, join_streamlet)
     self._add_child(join_streamlet)
@@ -190,7 +190,7 @@ class Streamlet(object):
     """Return a new Streamlet in which each (key, value) pair of this Streamlet are collected
        over the time_window and then reduced using the reduce_function
     """
-    from heronpy.dsl.reducebykeyandwindowbolt import ReduceByKeyAndWindowStreamlet
+    from heronpy.dsl.impl.reducebykeyandwindowbolt import ReduceByKeyAndWindowStreamlet
     reduce_streamlet = ReduceByKeyAndWindowStreamlet(window_config, reduce_function, self)
     self._add_child(reduce_streamlet)
     return reduce_streamlet
